refactor: move `TagPage`
