+++ lib/view/page/tag/tag_page.dart
@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../i18n/strings.g.dart';
-import '../../model/account.dart';
-import '../../provider/api/post_notifier_provider.dart';
-import '../../provider/api/tag_notes_notifier_provider.dart';
-import '../widget/note_widget.dart';
-import '../widget/paginated_list_view.dart';
+import '../../../i18n/strings.g.dart';
+import '../../../model/account.dart';
+import '../../../provider/api/post_notifier_provider.dart';
+import '../../../provider/api/tag_notes_notifier_provider.dart';
+import '../../widget/note_widget.dart';
+import '../../widget/paginated_list_view.dart';
 
 class TagPage extends ConsumerWidget {
   const TagPage({
